Revert "Add action for running tests on push"
This reverts commit 8de512c4057782269edd20d1466068b2633166f7.
The reason is that github CI environment fails to run tests that include
the run() function.
Strangely, this happens also if Sensor is not used, while I would have
guessed that Sensor was the culprit.
Anyway, I do not have time to fix this and do not need actions now.